requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
pydantic>=2.7
python-dotenv>=1.0

torch>=2.2           # install CUDA build if you have GPU (see note below)
transformers>=4.43
accelerate>=0.33
safetensors>=0.4.5

pdfplumber>=0.11
python-slugify>=8.0

sentence-transformers>=3.0
numpy>=1.24

rank-bm25>=0.2.2
nltk>=3.8.1

pyyaml>=6.0.1

prometheus-client>=0.20
opentelemetry-api>=1.26.0
opentelemetry-sdk>=1.26.0
opentelemetry-instrumentation-fastapi>=0.47b0

langgraph
langchain-core

groq

faiss-cpu>=1.8.0
qdrant-client>=1.9.2

pytest
fitz
pymupdf
pytest-cov>=4.0
ruff>=0.6

ddgs
